+++ src/audio/mod.rs
@@ -1,0 +1,7 @@
+pub mod frequencies;
+pub mod keys;
+pub mod oscillator;
+
+pub use frequencies::*;
+pub use keys::*;
+pub use oscillator::*;
