+++ src/dsp/signal_processing.rs
@@ -1,6 +1,6 @@
 use libm::{expf, logf};
 
-use crate::{MusicalSettings, fft::FftOps};
+use crate::{MusicalSettings, dsp::FftOps};
 
 /// Extract cepstral envelope for formant preservation using generic FFT operations
 pub fn extract_cepstral_envelope<const N: usize, const HALF_N: usize, F>(
@@ -51,15 +51,18 @@ pub fn calculate_pitch_shift(
 ) -> f32 {
     let mut pitch_shift_ratio = previous_pitch_shift_ratio;
     let fundamental_index =
-        crate::process_frequencies::find_fundamental_frequency(analysis_magnitudes);
+        crate::dsp::frequency_analysis::find_fundamental_frequency(analysis_magnitudes);
     let detected_frequency = analysis_frequencies[fundamental_index] * bin_width;
 
     if detected_frequency > 0.001 {
         let target_frequency = if settings.note == 0 {
-            let scale_frequencies = crate::keys::get_scale_by_key(settings.key);
-            crate::frequencies::find_nearest_note_in_key(detected_frequency, scale_frequencies)
+            let scale_frequencies = crate::audio::keys::get_scale_by_key(settings.key);
+            crate::audio::frequencies::find_nearest_note_in_key(
+                detected_frequency,
+                scale_frequencies,
+            )
         } else {
-            crate::keys::get_frequency(settings.key, settings.note, settings.octave, false)
+            crate::audio::keys::get_frequency(settings.key, settings.note, settings.octave, false)
         };
         let raw_ratio = target_frequency / detected_frequency;
         let clamped_ratio = raw_ratio.clamp(0.5, 2.0);
