--- src/utils/mod.rs
@@ -1,8 +1,0 @@
-//! Utility modules for vocal effects processing
-
-pub mod math;
-pub mod signal;
-
-// Re-export commonly used utilities
-pub use math::*;
-pub use signal::*;
